compact: Fixed populateBlock to fix out ouf boundary chunks mautomatically.
Fixes https://github.com/prometheus/prometheus/issues/5105

This is seems to be in the case when you do snapshot and it will invoke `write.Compact` for head
block with certain MaxTime being time.Now in the moment of invoke. During compaction
head block's chunks can be appended anytime resulting in chunks being out max Time.

Changes:
* For block write:
    * If chunk is partially outside max time, remove leftovers
    * If chunk is completely outside skip it.
    * If chunk is before min time - error out, we don't have use case for this.
* Improvement comments for compactor write API. We aim to guarantee block chunks being
within requested min and max time.
* Added test cases
* Avoided returning via argument if not needed (meta stats)
* Handled error from reader closers in populate Block. For now they are logged as error level.
(We should not block write if we cannot close reader I guess, but we have to be aware at least)

Just to double check: Blocks has (minTime, maxTime> chunks right? Where was documentation for this?
I remember we wanted to note down this somewhere (:

Signed-off-by: Bartek Plotka <bwplotka@gmail.com>